perf(frontend): enable Streaming SSR for homepage with Suspense boundary
Move heavy data fetching (properties + translations) out of page.tsx
into a child async Server Component (List.Streamed), wrapped in Suspense.
Shell (AnimatedText, CategoryScroller, skeleton) now streams immediately
instead of waiting for backend API + Google Translate to resolve.

Made-with: Cursor
